Applying Rules
Learning how to apply rules and keep them in reference. The rule does not yet apply; the goal is to get the generated GameStateDelta to be a source of what will actually happen. Passing the "ScriptingExecutionContext" back seems like the obvious thing to do... but maybe we need to cut that object in to a bit of a smaller shape.