Teaching GameStateDeltaMaker to "trigger"
Midway progress on teaching GameStateDeltaMaker to handle triggers. A "Reduce" scripting command wants to see the stat trigger, with a pointer to it, so that the command can reduce the intensity. However, we've backed ourself into the ironic corner of needing to update GameStateDelta with all of the functions of GameState.