Representing Card Target Question
When selecting a card to play, it hooks up with the old target highlighting system. Now the you can see the valid targets for the card, although it doesn't respond to being clicked on yet.

Lots of changes because the `ExecutionQuestion`, `ExecutionAnswer`, `ExecutionAnswerSet`, and `QuestionAnsweringContext` were moved to `GameState.Minimum`. This is to support `CardInstance` having a reference to it. Lots of small namespace tweaks.